lib: Support lazy initialization of per-cpu counters
While per-cpu counters are very efficient when there is a need for
frequent updates from different cpus, they have a non-trivial upfront
initialization cost, mainly due to the percpu variable allocation.  This
cost becomes relevant either for short-lived counters and for cases
where we don't know beforehand if there will be frequent updates from
remote cpus. On both cases, it could have been better to just use a
simple counter.

The prime example is rss_stats of single-threaded tasks, where the vast
majority of counter updates happen from a single-cpu context at a time.
For those workloads, a simple counter would have sufficed and likely
yielded better overall performance if the tasks are sufficient short.
There is no end of examples of short-lived single-thread benchmarks, one
example is the kernel compilation itself.

This patch introduces a new counter flavor that delays the percpu
initialization until needed.  It is a dual-mode counter.  It starts with
a two-part counter that can be updated either from a local context
through simple arithmetic or from a remote context through an atomic
operation.  Once remote accesses become more frequent, and the user
considers the overhead of atomic updates surpasses the cost of
initializing a fully-fledged per-cpu counter, the user can seamlessly
upgrade the counter to the per-cpu counter.

Features:
  0) Seamless migration to full percpu_counter
  1) Doesn't consume more space than a struct percpu_counter
  2) Allow atomic updates with a fast path for local updates
  4) The error of a remote reader can be bounded to a constant limit
  defined for that specific counter.
  5) batch initialization of an array of counters.

The first user of this are the rss_stat counters.  Benchmarks provided
on that patch.

Signed-off-by: Gabriel Krisman Bertazi <krisman@suse.de>