mm: Avoid percpu MM counters on single-threaded tasks
The cost of the pcpu memory allocation when forking a new task is
non-negligible, as reported in a few occasions[1,2].

It can also be fully avoided for single-threaded applications, where we
know the vast majority of updates happen from the local task context.

For the trivial benchmark bound to cpu 0 (to reduce cost of migrations),
like below:

     for (( i = 0; i < 20000; i++ )); do /bin/true; done

on an 80c machine, this patchset yielded a 6% improvement in elapsed
time.  On a 256c machine, the elapsed time reduced beyond 15%. Profiling
shows mm_init went from 13.5% of samples to less than 3.33% on the same
256c machine:

Before:
-   13.50%     3.93%  benchmark.sh     [kernel.kallsyms] [k] mm_init
   - 9.57% mm_init
      + 4.80% pcpu_alloc_noprof
      + 3.87% __percpu_counter_init_many

After:
-    3.33%     0.80%  benchmark.sh  [kernel.kallsyms]  [k] mm_init
   - 2.53% mm_init
      + 2.05% pcpu_alloc_noprof

For kernbench on the same machine, it still yielded 1.4%
improvement on system time.